Add setuptools version guard to setup.py.
Fully embrancing setup.cfg requires a newish version of setuptools.
Unfortunately there is not a great way to tell the user when they need a
newer version.  The most reliable mechanism that I've come up with is to
use pkg_resources to compare the versions and explode if it is too old.